Create a ConsentRepository interface and move the implementation to the bundle
