+++ src/OpenConext/EngineBlock/Service/ConsentService.php
@@ -5,9 +5,9 @@ namespace OpenConext\EngineBlock\Service;
 use Doctrine\DBAL\DBALException;
 use OpenConext\EngineBlock\Authentication\Dto\Consent;
 use OpenConext\EngineBlock\Authentication\Dto\ConsentList;
-use OpenConext\EngineBlock\Exception\RuntimeException;
 use OpenConext\EngineBlock\Authentication\Model\Consent as ConsentEntity;
 use OpenConext\EngineBlock\Authentication\Repository\ConsentRepository;
+use OpenConext\EngineBlock\Exception\RuntimeException;
 use OpenConext\Value\Saml\EntityId;
 use Psr\Log\LoggerInterface;
